move to git ... oops!!
sorry if you're reading this looking for more commits, or anything presentable!
I started this project at the end of 2017 intending it just as a fun hobby for
myself, so I didn't bother to set up a repository, and put off detailed
commenting until I was happy with everything else, which never quite happened.
I took a break from it for a while but I want to try again now! and maybe tidy
up a bit.