a lemma about Enumerations
a list with unique entries that's the same length as an Enumeration of a type
contains all elements of that type. this lets us know that, for example, if we
have a safe tile with n adjacent mines, and a list of n mines neighboring it,
then that list is complete.