add a rule for case analysis
the previous rules for constructing proofs that minesweeper moves are valid
weren't strong enough to express all valid moves. for example, on the board
```
**
*6 1
**```
if the tiles marked with `*`s are mines, all of the tiles adjacent to the `1`
but not the `6` will be safe. however, before, this could not be proven, since
we do not know the exact positions of all the mines or safe tiles adjacent to
either the `6` or `1`, and that was necessary information to make any judgment.
the new rule captures the reasoning that the `1`'s neighbors not adjacent to the
`6` will always be safe, since they can be shown to be safe for any way of
picking which of the `6`'s unknown neighbors is the mine,