rewrite Enumeration
now, instead of defining Enumeration--a collection of all unique elements of a
type--as a list of all of its elements (unique up to propositional equality),
it is a bijection between a finite set and its elements (up to a provided
equivalence). mostly, I think this makes things nicer! it allows for expressing
more things in terms of standard library functions, and simplifies a few
others, but some did need extra work to transfer over, and overall it was a big
rewrite. I want this to look as presentable as I can make it, though, so I think
it's worth it.